Add files via upload
This is the solar calculator class used to do all the calculations. 

Create a new instance of SolCalc 

call SolCalc.GetTotalSystemCost()

it expects three parameters : 

JPS monthly bill : decimal 
percentage to cut by: decimal ( in the form 0.XX) 
and a boolean which represents "Batteries included"

It return a quote object which contains all the information on the solar quote. 

You can access the info in this format : 

   val calc = SolCalc()
   println(calc.GetTotalSystemCost(45000.0,.90,true).totalInstallationCost)

Look in the class quote class to see all the information that's available.